fix: Simplify fetching
