+++ src/components.ts
@@ -1,5 +1,4 @@
 import path from "path"
-import * as nodeFetch from "node-fetch"
 import { createDotEnvConfigComponent } from "@well-known-components/env-config-provider"
 import { createServerComponent, createStatusCheckComponent } from "@well-known-components/http-server"
 import { createLogComponent } from "@well-known-components/logger"
@@ -9,12 +8,11 @@ import { createTracerComponent } from "@well-known-components/tracer-component"
 import { createHttpTracerComponent } from "@well-known-components/http-tracer-component"
 import { createMetricsComponent } from "@well-known-components/metrics"
 import { AppComponents, GlobalContext } from "./types"
-import { createFetchComponent } from "./ports/fetch/fetch"
+import { createFetchComponent } from "./ports/fetch"
 import { metricDeclarations } from "./metrics"
 import { createSchemaValidatorComponent } from "./ports/schema-validator"
 import { createRentalsComponent } from "./ports/rentals/component"
 import { createJobComponent } from "./ports/job/component"
-import { RequestInterceptor } from "./ports/fetch"
 
 // Initialize all the components of the app
 export async function initComponents(): Promise<AppComponents> {
@@ -33,23 +31,9 @@ export async function initComponents(): Promise<AppComponents> {
   const logs = await createLogComponent({ tracer })
   const server = await createServerComponent<GlobalContext>({ config, logs }, { cors })
   createHttpTracerComponent({ server, tracer })
-
-  const traceRequestInterceptor: RequestInterceptor = (url, init) => {
-    const headers: nodeFetch.HeadersInit = { ...init?.headers }
-    const traceParent = tracer.isInsideOfTraceSpan() ? tracer.getTraceChildString() : null
-    if (traceParent) {
-      ;(headers as { [key: string]: string }).traceparent = traceParent
-      const traceState = tracer.getTraceStateString()
-      if (traceState) {
-        ;(headers as { [key: string]: string }).tracestate = traceState
-      }
-    }
-
-    return [url, { ...init, headers }]
-  }
   const statusChecks = await createStatusCheckComponent({ server, config })
 
-  const fetch = createFetchComponent(nodeFetch.default, { requestInterceptors: [traceRequestInterceptor] })
+  const fetch = await createFetchComponent({ tracer })
   const metrics = await createMetricsComponent(metricDeclarations, { server, config })
   const marketplaceSubgraph = await createSubgraphComponent({ logs, config, fetch, metrics }, MARKETPLACE_SUBGRAPH_URL)
   const rentalsSubgraph = await createSubgraphComponent({ logs, config, fetch, metrics }, RENTALS_SUBGRAPH_URL)
