--- src/ports/fetch/index.ts
@@ -1,2 +1,0 @@
-export * from "./fetch"
-export * from "./types"
